+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## v0.2.1
 
+### Portfolio Improvements
+
+- Add performance metrics to `Portfolio`
+
 ### Indicator Improvements
 
 - Add `Err` return values to `BBands.process_graph_for_new_candles()`
-- Formally test `BBands.process_graph_for_new_candles()`
+- Formally test `BBands.Gprocess_graph_for_new_candles()`
 
 ---